style: Refactor styles
