+++ global.d.ts
@@ -1,0 +1,9 @@
+declare module '*.module.css' {
+  const content: { [className: string]: string };
+  export default content;
+}
+
+declare module '*.css' {
+  const content: string;
+  export default content;
+}
