--- src/app/App.css
@@ -1,7 +1,0 @@
-#root {
-  display: flex;
-  margin: 0 auto;
-  padding: 1rem;
-  text-align: center;
-  min-height: 100vh;
-}
