--- src/app/App.tsx-
@@ -1,26 +1,0 @@
-import './App.css';
-import { Wrapper } from 'src/components/wrapper/wrapper';
-
-import { Route, Routes } from 'react-router';
-import HomePage from '@/app/pages/HomePage/HomePage';
-import About from '@/app/pages/About/About';
-import Page404 from '@/app/pages/Page404/Page404';
-import Header from 'src/components/header/header';
-import { Details } from 'src/components/results/details/details';
-
-const App = () => {
-  return (
-    <Wrapper>
-      <Header />
-      <Routes>
-        <Route path="/" element={<HomePage />}>
-          <Route path="details" element={<Details />} />
-        </Route>
-        <Route path="about" element={<About />} />
-        <Route path="*" element={<Page404 />} />
-      </Routes>
-    </Wrapper>
-  );
-};
-
-export default App;
