+++ src/app/globals.css
@@ -10,6 +10,14 @@
   font-weight: 400;
 }
 
+#root {
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+  text-align: center;
+  min-height: 100vh;
+}
+
 body {
   background-color: var(--background);
   color: var(--text);
@@ -22,7 +30,7 @@ body {
   --table-row-hover: #5f9ea0;
   --table-row-selected: #a9a9a9;
   --popup-background: #6c6c6ce6;
-  --url-404: url('./404-dark.jpg');
+  --url-404: url('/images/404-dark.jpg');
   --background-disabled: #353131;
 }
 
@@ -33,7 +41,7 @@ body {
   --table-row-hover: #89dde0;
   --table-row-selected: #c7eb63;
   --popup-background: #cecdcde6;
-  --url-404: url('./404-light.jpg');
+  --url-404: url('/images/404-light.jpg');
   --background-disabled: #5a5a5a;
 }
