+++ src/app/layout.tsx
@@ -1,5 +1,7 @@
 'use client';
 
+import './globals.css';
+
 import Header from 'src/components/header/header';
 import { ThemeProvider } from 'src/context/themeProvider';
 import { Wrapper } from '@/wrapper';
@@ -15,7 +17,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
-        <link rel="icon" type="image/svg+xml" href="/src/assets/react.svg" />
+        <link rel="icon" type="image/svg+xml" href="./images/react.svg" />
         <title>Star Trek API. Next SSR</title>
       </head>
       <body>
