+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import Page404 from './Page404';
+import Page404 from './page404/page';
 
 export default function NotFound() {
   return <Page404 />;
