+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from './HomePage';
+import HomePage from './homePage/page';
 
 export default function NotFound() {
   return <HomePage />;
